Create Player Script
-remove playerMovement, playerHealth, playerShoot
-put them together in a Player script, most variables are in the Start
function for easy manipulation
-remove HP dependent codes in other scripts that are not necessary
-added a player2 for second playable character, that is slower, bulkier,
fire more bullets at once
