+++ Assets/scripts/Player.cs.meta
@@ -1,7 +1,7 @@
 fileFormatVersion: 2
-guid: d224ce41a714ef64184fc74d46e1a76d
-timeCreated: 1452922845
-licenseType: Free
+guid: 41ed54eaba5af904aa4835c6d424d8d8
+timeCreated: 1455228529
+licenseType: Pro
 MonoImporter:
   serializedVersion: 2
   defaultReferences: []
