+++ Assets/scripts/Player2.cs.meta
@@ -1,7 +1,7 @@
 fileFormatVersion: 2
-guid: 9bc8c6ffb5ea75a458512340f308fcfe
-timeCreated: 1454505824
-licenseType: Free
+guid: 3973d8ee3caa49c48bbb89a818e33513
+timeCreated: 1455231305
+licenseType: Pro
 MonoImporter:
   serializedVersion: 2
   defaultReferences: []
