+++ Assets/scripts/itemCollide.cs
@@ -6,7 +6,7 @@ public class itemCollide : MonoBehaviour {
 	void OnTriggerEnter2D (Collider2D coll) {
 		if (coll.gameObject.tag == "player") {
 			Destroy(gameObject);
-			playerShoot.itemBoost = true;
+			Player.itemBoost = true;
 		}
 	}
 }
