--- Assets/scripts/playerMovement.cs
@@ -1,44 +1,0 @@
-﻿using UnityEngine;
-using System.Collections;
-
-public class playerMovement : MonoBehaviour {
-    
-   
-    float speed = 5.0f;  //Default speed for player.
-    float subBoundaryRadius = 1f;   //Border for the submarine. used for movement restriction
-
-
-    void Update () {
-
-        float screenRatio = (float)Screen.width / (float)Screen.height;
-        float widthOrtho = Camera.main.orthographicSize * screenRatio;
-
-        Vector3 pos = transform.position;
-        /*PLAYER MOVEMENT    
-            GetAxisRaw inputs can be changed in Edit > Project Settings > Input. Disabled alternative keys for simplicity. 
-            May have alternative buttons later on.
-        */
-        pos.y += Input.GetAxisRaw("Vertical") * speed * Time.deltaTime;
-        pos.x += Input.GetAxisRaw("Horizontal") * speed * Time.deltaTime;
-       
-        /*RESTRICT MOVEMENT WITHIN MAIN CAMERA
-            Checks submarine position + the border of it to see if its near the edges of the main camera.
-            OrthographicSize takes vertical positions only therefore widthOrtho is made to define the horizontal position of the camera.
-            Vertical position may change due to implementation of UI.
-        */
-        if(pos.y + subBoundaryRadius > Camera.main.orthographicSize){
-            pos.y = Camera.main.orthographicSize - subBoundaryRadius;
-        }
-        if(pos.y - subBoundaryRadius < -Camera.main.orthographicSize){
-            pos.y = -Camera.main.orthographicSize + subBoundaryRadius;
-        }   
-        if(pos.x + subBoundaryRadius > widthOrtho){
-            pos.x = widthOrtho - subBoundaryRadius;
-        }
-        if (pos.x - subBoundaryRadius < -widthOrtho){
-            pos.x = -widthOrtho + subBoundaryRadius;
-        }
-        //Updates player position.
-        transform.position = pos;
-    }
-}
