--- Assets/scripts/playerMovement.cs.meta
@@ -1,12 +1,0 @@
-fileFormatVersion: 2
-guid: eb8dce116f16b3742b95576b09c28d1a
-timeCreated: 1452915137
-licenseType: Free
-MonoImporter:
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
